+++ src/OmnichannelChatSDK.ts
@@ -109,6 +109,7 @@ import startPolling from "./commands/startPolling";
 import stopPolling from "./commands/stopPolling";
 import urlResolvers from "./utils/urlResolvers";
 import validateOmnichannelConfig from "./validators/OmnichannelConfigValidator";
+import FetchChatTokenResponse from "@microsoft/ocsdk/lib/Model/FetchChatTokenResponse";
 
 class OmnichannelChatSDK {
     private debug: boolean;
@@ -639,11 +640,12 @@ class OmnichannelChatSDK {
             }
         }
 
-        if (this.chatToken && Object.keys(this.chatToken).length === 0) {
+        if (this.chatSDKConfig.useCreateConversation?.disable && this.chatToken && Object.keys(this.chatToken).length === 0) {
             await this.getChatToken(false);
         }
-
-        loggerUtils.setChatId(this.chatToken.chatId || '', this.ocSdkLogger, this.acsClientLogger, this.acsAdapterLogger, this.callingSdkLogger, this.amsClientLogger, this.ic3ClientLogger);
+        if (this.chatToken?.chatId) {
+            loggerUtils.setChatId(this.chatToken.chatId || '', this.ocSdkLogger, this.acsClientLogger, this.acsAdapterLogger, this.callingSdkLogger, this.amsClientLogger, this.ic3ClientLogger);
+        }
 
         let sessionInitOptionalParams: ISessionInitOptionalParams = {
             initContext: {} as InitContext
@@ -677,6 +679,31 @@ class OmnichannelChatSDK {
             if (!optionalParams.liveChatContext) {
                 try {
                     await this.OCClient.sessionInit(this.requestId, sessionInitOptionalParams);
+                } catch (error) {
+                    const telemetryData = {
+                        RequestId: this.requestId,
+                        ChatId: this.chatToken.chatId as string,
+                    };
+
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    if ((error as any)?.isAxiosError && (error as any).response?.headers?.errorcode?.toString() === OmnichannelErrorCodes.WidgetUseOutsideOperatingHour.toString()) {
+                        exceptionThrowers.throwWidgetUseOutsideOperatingHour(error, this.scenarioMarker, TelemetryEvent.StartChat, telemetryData);
+                    }
+
+                    exceptionThrowers.throwConversationInitializationFailure(error, this.scenarioMarker, TelemetryEvent.StartChat, telemetryData);
+                }
+            }
+        };
+
+        const createConversationPromise = async () => {
+            // Skip session init when there's a valid live chat context
+            if (!optionalParams.liveChatContext) {
+                try {
+                    const chatToken = await this.OCClient.createConversation(this.requestId, sessionInitOptionalParams);
+                    if (chatToken) {
+                        this.setChatToken(chatToken);
+                        loggerUtils.setChatId(this.chatToken.chatId || '', this.ocSdkLogger, this.acsClientLogger, this.acsAdapterLogger, this.callingSdkLogger, this.amsClientLogger, this.ic3ClientLogger);
+                    }
                 } catch (error) {
                     const telemetryData = {
                         RequestId: this.requestId,
@@ -792,7 +819,12 @@ class OmnichannelChatSDK {
             }
         };
 
-        await Promise.all([sessionInitPromise(), messagingClientPromise(), attachmentClientPromise()]);
+        if (!this.chatSDKConfig.useCreateConversation?.disable) {
+            await createConversationPromise();
+            await Promise.all([messagingClientPromise(), attachmentClientPromise()]);
+        } else {
+            await Promise.all([sessionInitPromise(), messagingClientPromise(), attachmentClientPromise()]);
+        }
 
         if (this.isPersistentChat && !this.chatSDKConfig.persistentChat?.disable) {
             this.refreshTokenTimer = setInterval(async () => {
@@ -1082,26 +1114,7 @@ class OmnichannelChatSDK {
                 }
 
                 const chatToken = await this.OCClient.getChatToken(this.requestId, getChatTokenOptionalParams);
-                const { ChatId: chatId, Token: token, RegionGtms: regionGtms, ExpiresIn: expiresIn, VisitorId: visitorId, VoiceVideoCallToken: voiceVideoCallToken, ACSEndpoint: acsEndpoint, AttachmentConfiguration: attachmentConfiguration } = chatToken;
-                this.chatToken = {
-                    chatId,
-                    regionGTMS: JSON.parse(regionGtms),
-                    requestId: this.requestId,
-                    token,
-                    expiresIn,
-                    visitorId,
-                    voiceVideoCallToken,
-                    acsEndpoint,
-                };
-
-                if (attachmentConfiguration && attachmentConfiguration.AttachmentServiceEndpoint) {
-                    this.chatToken.amsEndpoint = attachmentConfiguration.AttachmentServiceEndpoint;
-                }
-
-                if (this.OCClient.sessionId) {
-                    this.sessionId = this.OCClient.sessionId;
-                }
-
+                this.setChatToken(chatToken)
                 this.scenarioMarker.completeScenario(TelemetryEvent.GetChatToken, {
                     RequestId: this.requestId,
                     ChatId: this.chatToken.chatId as string
@@ -1126,6 +1139,28 @@ class OmnichannelChatSDK {
         }
 
         return this.chatToken;
+    }
+
+    public setChatToken(chatToken: FetchChatTokenResponse): void {
+        const { ChatId: chatId, Token: token, RegionGtms: regionGtms, ExpiresIn: expiresIn, VisitorId: visitorId, VoiceVideoCallToken: voiceVideoCallToken, ACSEndpoint: acsEndpoint, AttachmentConfiguration: attachmentConfiguration } = chatToken;
+        this.chatToken = {
+            chatId,
+            regionGTMS: JSON.parse(regionGtms as string),
+            requestId: this.requestId,
+            token,
+            expiresIn,
+            visitorId,
+            voiceVideoCallToken,
+            acsEndpoint,
+        };
+
+        if (attachmentConfiguration && attachmentConfiguration.AttachmentServiceEndpoint) {
+            this.chatToken.amsEndpoint = attachmentConfiguration.AttachmentServiceEndpoint;
+        }
+
+        if (this.OCClient.sessionId) {
+            this.sessionId = this.OCClient.sessionId;
+        }
     }
 
     public async getCallingToken(): Promise<string> {